Implementados todos los filtros de búsqueda

+++ src/main/java/com/example/korner/modelo/AbstractEntity.java
@@ -30,7 +30,7 @@ public abstract class AbstractEntity implements Serializable {
 
     @Column (name = "year")
     @NotNull
-    private Short year;
+    private Integer year;
 
     @Column (name = "puntuacion")
     @Min(0)
